refactor: [#1157] drive LSP symbol index off TypedProgram + name_type_map
Delete `lsp/symbols.rs` (796 lines) and replace it with `index.rs` that
walks the already-typed `TypedProgram` instead of re-parsing. The
checker's new `name_type_map` (`name -> Arc<Type>`) feeds pipe-compat
ranking without per-keystroke `.to_string()` round-trips.

- Drop `enrich_hover_detail`; bake its logic into `SymbolIndex::build`
  so `Symbol.detail` is the final hover string.
- `enrich_from_imports` now emits the hover-ready import detail format
  directly; no post-hoc rendering.
- `Symbol.first_param_type` is `Option<Arc<Type>>`, typed pipe compat
  via `is_pipe_compatible_typed` avoids the old `.to_string()` alloc in
  the per-doc/per-symbol hot loop.
- Expose `checker::simple_resolve_type_expr` so the for-block path
  stops hand-rolling a syntactic->Type converter.
